Fix "flask run" not working without FLASK_APP
Previously we were trying to use main.py to run Celery and run Flask,
but these have different needs. This commit makes the way we run Flask
and the way we run Celery consistent with our other repos [1].

Resolves https://github.com/alphagov/notifications-govuk-alerts/pull/151#discussion_r687871510

[1]: https://github.com/alphagov/notifications-template-preview/blob/master/run_celery.py
