+++ run_celery.py
@@ -1,0 +1,4 @@
+from app import create_app, notify_celery  # noqa
+
+application = create_app()
+application.app_context().push()
